+++ src/Authentication/TwoFactorAuth/Infrastructure/Factory/ContentModelFactoryInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\SecurityModule\Authentication\TwoFactorAuth\Infrastructure\Factory;
+
+use OxidEsales\Eshop\Application\Model\Content;
+
+interface ContentModelFactoryInterface
+{
+    public function create(): Content;
+}